% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vl_package.R
\docType{data}
\name{govlab_color}
\alias{govlab_color}
\title{Fahrenheit conversion}
\format{
An object of class \code{character} of length 7.
}
\usage{
govlab_color
}
\description{
Convert degrees Fahrenheit temperatures to degrees Celsius
}
\keyword{datasets}
